Should be back to where we were 1 year ago.
I started to make some changes but then decided to discard them. The
problem is that Git does not compare the contents of the excel files,
it just compares the dates. And since the date is todays date when I
revert a change, then it thinks that all the .txt files are too old. So
I regenerated the text files. The reassuring thing is that there are
not differences detected in the .txt files, so we should be like we
were before.